requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
deepspeech==0.7.1
future==0.18.2
numpy==1.17.0
pycparser==2.20
pydub==0.24.0
python-telegram-bot==12.7
scipy==1.4.1
six==1.14.0
tornado==6.0.4
